Bug 566498 - Convert to HashMap to benefit from JEP-180
This change is quite intense. Performance should improve in a workspace
with many packages and potential collisions as JEP-180 guarantees a
faster lookup from the map. The costs of memory is a little higher but
given today's machines this shouldn't be a real concern.

One thing that I noticed is String interning in JavaModelManager. It
might make sense to move the pool to a more central place if a
dependency on JavaModelManager is not wanted. Also, with G1GC and string
de-duplication this should no longer be a real concern.

Change-Id: I3fad108a9f1b65e4d917f76d0962573896aab4be